requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
asyncssh
aiofiles
loguru
rich
questionary
python-dotenv